Refactor of storyboardGraph.ts. Ran format and linter
